replace arguments of launch_executable from command_line and wait to args
