+++ source/os/operations.c
@@ -1,6 +1,8 @@
 #include "operations.h"
 #include "../core/io_wrap.h"
 
+#include <stdio.h>
+
 #ifdef _WIN32
 #    include <WinBase.h>
 
@@ -76,7 +78,64 @@ void clear_screen() {
     // NOLINTEND
 }
 
-bool launch_executable(const os_char *command_line, bool wait) {
+bool is_empty_str(os_char *str){
+    return !strcmp(str, "");
+}
+
+void extract_from_args(const struct args args, os_char **p_command_line, bool *p_wait) {
+    bool wait = true;
+    bool flag = false;
+    size_t len = 0;
+    if(!is_empty_str(args.argv[0])) {
+        len += strlen(args.argv[0]) + 2;
+    }
+    for(int i = 1; i < args.argc; ++i) {
+        if(is_empty_str(args.argv[i])) {
+            continue;
+        }
+        if(strcmp(args.argv[i], "&") != 0 || flag) {
+            len += strlen(args.argv[i]) + 3;
+        }
+        else
+        if(!flag) {
+            wait = false;
+            flag = true;
+        }
+    }
+
+    os_char *command_line = malloc((len + 1) * sizeof(os_char));
+    flag = false;
+    len = 0;
+    if(!is_empty_str(args.argv[0])) {
+        sprintf(command_line, "\"%s\"", args.argv[0]);
+        len += strlen(args.argv[0]) + 2;
+    }
+    for(int i = 1; i < args.argc; ++i) {
+        if(is_empty_str(args.argv[i])) {
+            continue;
+        }
+        if(strcmp(args.argv[i], "&") != 0 || flag) {
+            sprintf(command_line + len, " \"%s\"", args.argv[i]);
+            len += strlen(args.argv[i]) + 3;
+        }
+        else
+        if(!flag) {
+            flag = true;
+        }
+    }
+
+    command_line[len] = '\0';
+
+    *p_command_line = command_line;
+    *p_wait = wait;
+}
+
+bool launch_executable(const struct args args) {
+    os_char *command_line = NULL;
+    bool wait = true;
+
+    extract_from_args(args, &command_line, &wait);
+
     STARTUPINFO si;
     PROCESS_INFORMATION pi;
 
@@ -88,6 +147,8 @@ bool launch_executable(const os_char *command_line, bool wait) {
     os_char *tmp_command_line = malloc((len + 1) * sizeof(os_char));
     memcpy(tmp_command_line, command_line, len * sizeof(os_char));
     tmp_command_line[len] = '\0';
+
+    free(command_line);
 
     if(!CreateProcess(
            NULL,             // No module name (use command line)
