+++ source/os/operations.h
@@ -3,6 +3,7 @@
 #include <stdbool.h>
 
 #include "type.h"
+#include "../core/args.h"
 
 /*! @brief Fetch current working directory and copy them into `buffer` */
 void get_cwd(unsigned int buffer_size, os_char *buffer);
@@ -16,5 +17,5 @@ bool change_cwd(const os_char *new_dir);
 /*! @brief Clear the console */
 void clear_screen();
 
-/*! @brief Run the executable foreground */
-bool launch_executable(const os_char *command_line, bool wait);
+/*! @brief Launch the executable */
+bool launch_executable(const struct args args);
